Move new message func to controller
Moved the logic for adding a new message to the
database to inside the the db.js file.
The indexRouter now calls on the the
addNewMessage function when a new message is
added.
The controller handles this call, which then calls
the function within the database file to push
the new message onto the messages object